requirements.txt
flask
googlenews
newspaper3k
nltk